Refactored Schema to recon_config.py, Fixed fmt issues
