+++ src/databricks/labs/remorph/reconcile/connectors/snowflake.py
@@ -4,10 +4,9 @@ from pyspark.errors import PySparkException
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col
 
-from databricks.labs.remorph.config import Schema
 from databricks.labs.remorph.reconcile.connectors.data_source import DataSource
 from databricks.labs.remorph.reconcile.constants import SourceDriver, SourceType
-from databricks.labs.remorph.reconcile.recon_config import JdbcReaderOptions
+from databricks.labs.remorph.reconcile.recon_config import JdbcReaderOptions, Schema
 
 
 class SnowflakeDataSource(DataSource):
